+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # ChangeLog
 
-* fix loop for node reset
-* improve gizomo
+* kubjs node - eval function
+* register all models under the ldlib namespace by default
+* add codeeditor for better text editing
+* SelectNode doesn't require trigger link anymore